requirements.txt
Django==1.11
django-markdown-app==0.9.2
django-markdown-deux==1.0.5
requests==2.18.4
gunicorn==19.7.1
dj_database_url==0.4.2
whitenoise==3.3.0
psycopg2==2.7.3
django-mptt==0.8.7
